LeicaDMA: Objective turret working
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@1421 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/LeicaDMI/LeicaDMIModel.h
@@ -96,7 +96,7 @@ public:
    bool IsMethodAvailable(int methodId);
 
    int magnification_;
-   int NA_;
+   double NA_;
    int articleNumber_;
    std::vector<bool> methods_;
    std::string immersionType_;
